+++ imagefield/management/commands/process_imagefields.py
@@ -51,7 +51,7 @@ class Command(BaseCommand):
             if field.field_label not in self._fields:
                 continue
 
-            queryset = field.model._default_manager.all()
+            queryset = field.model._default_manager.exclude(**{field.name: ""})
             count = queryset.count()
             self.stdout.write(
                 "%s - %s objects - %s"