+++ imagefield/management/commands/process_imagefields.py
@@ -63,7 +63,9 @@ class Command(BaseCommand):
             )
             self.stdout.write("\r|%s| %s/%s" % (" " * 50, 0, count), ending="")
 
-            for index, instance in enumerate(queryset):
+            # Relatively low chunk_size to avoid slowness when having to load
+            # width and height for images when instantiating models.
+            for index, instance in enumerate(queryset.iterator(chunk_size=100)):
                 fieldfile = getattr(instance, field.name)
                 if fieldfile and fieldfile.name:
                     for key in field.formats: