+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Change log
 ~~~~~~~~~~~~~~~
 
 - Fixed a crash where unpickling image fields would fail.
+- Changed ``process_imagefields`` to skip exclude model instances with
+  an empty image field.
 
 
 `0.5`_ (2018-08-15)